+++ docs/about/release-notes.rst
@@ -47,6 +47,7 @@ Features
   Avoid huge memory use and slow performance in use of :py:func:`scipp.bin` and :py:func:`scipp.group` when only entire bins are combined (rather than splitting bins based on bin content coordinates).
   This avoids a number of cases where Python kernel used to crash since Scipp ran out of memory `#2827 <https://github.com/scipp/scipp/pull/2827>`_.
 * :py:func:`scipp.sum` now also accepts a sequence of dimension labels `#2827 <https://github.com/scipp/scipp/pull/2827>`_.
+* Added aliases for units to customize string formatting and support user defined units `#2855 <https://github.com/scipp/scipp/pull/2855>`_.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~